requirements.txt
pymongo==4.7.3
ratelimit==2.2.1
Requests==2.32.3
tqdm==4.66.4
pre-commit
jsonPagination==0.3.1
pytz==2024.1
pyfiglet===1.0.2
termcolor==2.4.0
colorama==0.4.6
schedule==1.2.2
python-dateutil==2.9.0
